backend/python/requirements.txt
# TouchCLI Agent Service Dependencies
# Phase 2: Backend Infrastructure

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Cache & Queue
redis==5.0.1
celery==5.3.4

# Agent Framework
langgraph==0.0.1
langchain==0.0.340
openai==1.3.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
pyjwt==2.8.1

# Environment
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.3.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: WebSocket support
websockets==12.0
python-socketio==5.10.0
